refactor: Changes layout of reflection tab, adds acknowledgments
I decided to post the reflection and acknowledgments directly on the page instead of having a link to them. I like the look of this better as it feels more like a portfolio and less like a link aggregation page.

The full reflection is still linked in the showcase tab.